Update examples to reflect recent changes

+++ examples/reparameterisations_example.py
@@ -22,7 +22,7 @@ class HalfGaussian(Model):
 
     def log_prior(self, x):
         """Log-prior"""
-        log_p = np.log(self.in_bounds(x))
+        log_p = np.log(self.in_bounds(x), dtype='float')
         for bounds in self.bounds.values():
             log_p -= np.log(bounds[1] - bounds[0])
         return log_p
